+++ web/src/router/index.js
@@ -7,7 +7,7 @@ import AboutView from '@/views/home/AboutView.vue'
 import LoginView from '@/views/auth/LoginView.vue'
 import AdminView from '@/views/admin/AdminView.vue'
 import RegisterView from '@/views/auth/RegisterView.vue'
-import PasswordRecoveryView from '@/views/auth/PasswordRecoveryView.vue'
+import ForgotPasswordView from '@/views/auth/ForgotPasswordView.vue'
 import adminRoutes from "./admin.router";
 import PasswordResetView from "../views/auth/PasswordResetView.vue";
 
@@ -41,10 +41,10 @@ const router = createRouter({
       component: LoginView,
     },
     {
-      path: '/password_recovery',
-      name: 'password_recovery',
+      path: '/forgot_password',
+      name: 'forgot_password',
       meta: { requiresAuth: false },
-      component: PasswordRecoveryView
+      component: ForgotPasswordView
     },
 
     {
