+++ web/src/views/auth/ForgotPasswordView.vue
